chore(RHINENG-15653): Cleanup SystemsTable
