+++ src/SmartComponents/SystemsTable/SystemsTable.test.js
@@ -1,31 +1,52 @@
-import { render, screen } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import TestWrapper from '@/Utilities/TestWrapper';
-import { SystemsTable } from './SystemsTableRest';
-jest.mock('@/Utilities/hooks/useAPIV2FeatureFlag', () => jest.fn(() => false));
+import { render, waitFor } from '@testing-library/react';
+import SystemsTableRest from './SystemsTable';
+import { useSystemBulkSelect } from './hooks.js';
+import TestWrapper from '@redhat-cloud-services/frontend-components-utilities/TestingUtils/JestUtils/TestWrapper';
+jest.mock('./hooks.js', () => ({
+  ...jest.requireActual('./hooks.js'),
+  useSystemBulkSelect: jest.fn(() => ({})),
+  useSystemsExport: jest.fn(() => ({})),
+}));
+jest.mock('@/Utilities/hooks/useAPIV2FeatureFlag', () => jest.fn(() => true));
+jest.mock('@redhat-cloud-services/frontend-components/Inventory', () => ({
+  InventoryTable: jest.fn(({ getEntities }) => {
+    getEntities(10, {
+      filters: {
+        hostnameOrId: 'test-name',
+        tagFilters: [],
+      },
+    });
+    return <div data-testid="inventory-mock-component">Inventory</div>;
+  }),
+}));
 
-describe('SystemsTable', () => {
-  it('returns an Inventory Table', () => {
+const mockProps = {
+  columns: ['test-column'],
+  enableExport: true,
+  policies: [],
+  onSelect: jest.fn(),
+  fetchApi: jest.fn(() => Promise.resolve({ data: [], meta: {} })),
+  fetchCustomOSes: jest.fn(),
+  ignoreOsMajorVersion: false,
+  defaultFilter: 'someFilter ~ test',
+};
+
+describe('SystemsTableRest', () => {
+  it('Should connect inventory with compliance filters so that full filters are passed to bulk selection', async () => {
     render(
       <TestWrapper>
-        <SystemsTable />
+        <SystemsTableRest {...mockProps} />
       </TestWrapper>
     );
-
-    expect(screen.getByLabelText('Inventory Table')).toBeInTheDocument();
-  });
-
-  it('returns an Inventory Table', () => {
-    render(
-      <TestWrapper>
-        <SystemsTable error={{ message: 'Error' }} />
-      </TestWrapper>
+    await waitFor(() =>
+      expect(useSystemBulkSelect).toHaveBeenCalledWith(
+        expect.objectContaining({
+          fetchArguments: {
+            filter: '(someFilter ~ test) AND display_name ~ "test-name"',
+            tags: [],
+          },
+        })
+      )
     );
-
-    expect(
-      screen.getByRole('heading', { name: 'Something went wrong' })
-    ).toBeInTheDocument();
   });
-
-  // TODO There should also be a test to verify the empty state
 });
